feat: 入力制限処理を追加
